Move Xml reader/writer base to fe-data for reuse
Easily reusable by other frontends that use XML.
